+++ src/app/(sok)/_components/filters/Extent.jsx
@@ -4,7 +4,7 @@ import { BodyShort, Checkbox, CheckboxGroup } from "@navikt/ds-react";
 import mergeCount from "@/app/(sok)/_components/utils/mergeCount";
 import { logFilterChanged } from "@/app/_common/monitoring/amplitude";
 import { EXTENT } from "@/app/(sok)/_components/searchParamNames";
-import useSearchQuery from "@/app/(sok)/_components/SearchStateProvider";
+import useSearchQuery from "@/app/(sok)/_components/SearchQueryProvider";
 
 function Extent({ initialValues, updatedValues }) {
     const values = mergeCount(initialValues, updatedValues);
